+++ CHANGELOG.md
@@ -1,2 +1,11 @@
+# v.0.2.0
+
+[BREAKING] - New macro, please replace.
+
+* Support for currency items
+* Support for random coin items
+* Support for random treasure table items.
+
 # v.0.1.0
+
 * Initial version for internal testing